backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
pydantic-settings
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
